Organized repo and changed makefile accordingly. Added options to select images to main and counter programs. Changed README.md accordingly.
